+++ src/components/markdown/index.js
@@ -1,6 +1,7 @@
 import CallSummaryTable from './call_summary_table'
 import ContactFormToggle from './contact_form_toggle'
 import HighlightCode from './highlight_code'
+import InlineImage from './inline_image'
 import Label from './label'
 
-export { CallSummaryTable, ContactFormToggle, HighlightCode, Label }
+export { CallSummaryTable, ContactFormToggle, HighlightCode, InlineImage, Label }
